% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcPenetrance.R
\name{mcmcPenetrance}
\alias{mcmcPenetrance}
\title{Bayesian Inference using Independent Metropolis-Hastings for Penetrance Estimation}
\usage{
mcmcPenetrance(
  data,
  n_chains,
  n_iter_per_chain,
  save_interval,
  m1,
  m2,
  max_age,
  shift_prior_min,
  shift_prior_max,
  p0,
  q1,
  q2,
  g1,
  g2
)
}
\arguments{
\item{data}{List of families data.}

\item{n_chains}{Number of chains for parallel computation.}

\item{n_iter_per_chain}{Number of iterations for each chain.}

\item{save_interval}{Interval after which states should be saved.}

\item{m1}{parameter for the beta distribution for the median.}

\item{m2}{parameter for the beta distribution fo the median.}

\item{max_age}{Maximum age to be considered.}

\item{shift_prior_min}{Minimum possible value for the shift parameter.}

\item{shift_prior_max}{Maximum possible value for the shift parameter.}

\item{p0}{baseline lifetime risk.}

\item{q1}{parameter of the beta distribution for the first quartile.}

\item{q2}{parameter of the beta distribution for the first quartile.}

\item{g1}{parameter of the beta distribution for the shift.}

\item{g2}{parameter of the beta distribution for the shift.}

\item{paras}{Vector of parameters for which likelihood is to be computed.}

\item{seed}{Seed value for random number generation.}

\item{n_iter}{Number of iterations for the chain.}

\item{chain_id}{Identifier for the chain.}
}
\value{
A list containing results for each chain.

Log-likelihood value.

A list with samples and rejection rate.
}
\description{
This function employs a Bayesian approach for penetrance estimation, utilizing the
Independent Metropolis-Hastings algorithm. It leverages parallel computing and requires
the \code{stats}, \code{parallel}, and \code{PPP} packages.
}
